Support C++ fast path for `DTensor.__torch_dispatch__`
This patches the `__torch_dispatch__` machinery to detect DTensor and hand over control to a C++ fast path. Unlike #166370 and #166369 (which added a DTensor dispatch key and are intended to be replaced by this PR), this approach fundamentally *is* `__torch_dispatch__`, hopefully sidestepping all manner of thorny "does it work just like `__torch_dispatch__`?" that came up during development and review of #166370.

ghstack-source-id: 74e4b634e6cd6d3829c5b26ee6e9a9a0460819d5
Pull Request resolved: https://github.com/pytorch/pytorch/pull/167051